requirements.txt
beautifulsoup4==4.7.1
certifi==2019.6.16
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1.4
django-js-asset==1.2.2
django-mptt==0.10.0
djangorestframework==3.9.1
djangorestframework-recursive==0.1.2
gunicorn==19.9.0
idna==2.8
psycopg2-binary==2.7.6.1
pytz==2018.7
requests==2.22.0
soupsieve==1.9.2
urllib3==1.25.3
whitenoise==4.0
